+++ app/src/main/java/woowacourse/shopping/data/dataSource/remote/CartProductService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.dataSource.remote.cart
+package woowacourse.shopping.data.dataSource.remote
 
 import retrofit2.Call
 import retrofit2.http.Body
